fix: move the navigation items to external file
